requirements.txt
# Core Libraries
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.2
seaborn==0.12.2

# Machine Learning Framework
torch==2.0.1  # For PyTorch
torchvision==0.20.1  # For image transforms and datasets

# Deployment Framework
flask==2.3.3
gunicorn==20.1.0  # Production server for Flask

# Additional Utilities
scikit-learn==1.3.1
Pillow==9.5.0  # Image processing
tqdm==4.66.0  # Progress bar
optuna==4.1.0

# Optional for Streamlit (if needed for visualization)
streamlit==1.26.0